update init file
